refactor(xcm): add pallet_message_queue into xcm config module
